Finally working properly
The rotational constants are finally being computed correctly, and the
correct rotation matrix is now being applied. Previously, I just used
the eigenvectors to rotate the molecule but that appeared to be slightly
off. Now we're using SVD to compute the rotation matrix of the inertial
tensor, which is then applied to the coordinates.

There was another incorrect behavior for the symmetrization of the
inertia tensor: `np.maximum` is not strictly correct for whatever
reason, whereas now I'm explicitly copying the matrix elements to
symmetrize. These problems were only apparent when the number of
isotopologues coming up were way more than what we would expect with
symmetry.